# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo produto",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os produtos",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do produto",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar um produto por ID",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do produto",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar um produto existente",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do produto",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto excluído com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir um produto",
        "tags": [
          "products"
        ]
      }
    },
    "/sectors": {
      "post": {
        "operationId": "SectorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Setor criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo setor",
        "tags": [
          "sectors"
        ]
      },
      "get": {
        "operationId": "SectorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de setores retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os setores",
        "tags": [
          "sectors"
        ]
      }
    },
    "/sectors/{id}": {
      "get": {
        "operationId": "SectorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do setor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setor encontrado."
          },
          "404": {
            "description": "Setor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter um setor pelo ID",
        "tags": [
          "sectors"
        ]
      },
      "put": {
        "operationId": "SectorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do setor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSectorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setor atualizado com sucesso."
          },
          "404": {
            "description": "Setor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar um setor pelo ID",
        "tags": [
          "sectors"
        ]
      },
      "delete": {
        "operationId": "SectorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do setor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setor removido com sucesso."
          },
          "404": {
            "description": "Setor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover um setor pelo ID",
        "tags": [
          "sectors"
        ]
      }
    },
    "/sectors/{id}/config": {
      "post": {
        "operationId": "SectorsController_addConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do setor",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSectorConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Configuração adicionada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar uma configuração personalizada a um setor",
        "tags": [
          "sectors"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "API documentation for NestJS with Swagger, Orval, and MikroORM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do produto",
            "example": "PROD001"
          },
          "name": {
            "type": "string",
            "description": "Nome do produto",
            "example": "Produto A"
          },
          "description": {
            "type": "string",
            "description": "Descrição do produto",
            "example": "Descrição detalhada do produto"
          },
          "price": {
            "type": "number",
            "description": "Preço do produto",
            "example": 99.99
          },
          "quantity": {
            "type": "number",
            "description": "Quantidade em estoque",
            "example": 100
          }
        },
        "required": [
          "code",
          "name",
          "price",
          "quantity"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do produto",
            "example": "PROD001"
          },
          "name": {
            "type": "string",
            "description": "Nome do produto",
            "example": "Produto A"
          },
          "description": {
            "type": "string",
            "description": "Descrição do produto",
            "example": "Descrição detalhada do produto"
          },
          "price": {
            "type": "number",
            "description": "Preço do produto",
            "example": 99.99
          },
          "quantity": {
            "type": "number",
            "description": "Quantidade em estoque",
            "example": 100
          }
        }
      },
      "CreateSectorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do setor",
            "example": "Usinagem"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateSectorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do setor",
            "example": "Usinagem e Montagem"
          }
        }
      },
      "AddSectorConfigDto": {
        "type": "object",
        "properties": {
          "fieldName": {
            "type": "string",
            "description": "Nome do campo de configuração",
            "example": "Temperatura Máxima"
          },
          "fieldType": {
            "type": "string",
            "description": "Tipo do campo de configuração",
            "example": "number"
          }
        },
        "required": [
          "fieldName",
          "fieldType"
        ]
      },
...
    }
  }
}